+++ allianceutils/__init__.py
@@ -1,9 +1,5 @@
 import unittest
 
-import django
-
-if django.VERSION < (3, 2):
-  default_app_config = 'allianceutils.apps.AllianceUtilsAppConfig'
 
 # All the tests are in test_allianceutils
 # Django's test autodiscovery will try to recursively import everything which causes asynctask to fail if we're running